chore: move `toggleState()` helper to general utils and export from lib
